+++ src/Fantomas.Tests/QuotationTests.fs
@@ -18,8 +18,10 @@ let ``typed quotations`` () =
     |> should
         equal
         """
-<@ let f x = x + 10
-   f 20 @>
+<@
+    let f x = x + 10
+    f 20
+@>
 """
 
 [<Test>]
@@ -39,4 +41,39 @@ let logger =
         .Setup(fun log -> <@ log.Log(error) @>)
         .Returns(())
         .Create()
+"""
+
+[<Test>]
+let ``should format multiline quotation expressions idempotent, 2203`` () =
+    formatSourceString
+        false
+        """
+let action =
+    <@
+        let msg = %httpRequestMessageWithPayload
+        RuntimeHelpers.fillHeaders msg %heads
+        async {
+            let! response = (%this).HttpClient.SendAsync(msg) |> Async.AwaitTask
+            return response.EnsureSuccessStatusCode().Content
+        }
+    @>
+"""
+        config
+    |> prepend newline
+    |> should
+        equal
+        """
+let action =
+    <@
+        let msg = %httpRequestMessageWithPayload
+        RuntimeHelpers.fillHeaders msg %heads
+
+        async {
+            let! response =
+                (%this).HttpClient.SendAsync(msg)
+                |> Async.AwaitTask
+
+            return response.EnsureSuccessStatusCode().Content
+        }
+    @>
 """